Several Linux and Windows enhancements.
- On all platforms, generate a sh header at the top of the module, so
it can be directly executed, in case it is built with a main program.
This depends on finding 'tclsh' on the path.

- On Windows, generate a .bat file adaptor to go alongside the module,
so it can be directly executed on Windows. This depends on finding
'tclsh' on the path.

- On Windows, default VCPKG configuration to build x64-windows-static
libraries, because it's easier to statically link everything into the
Tcl extension.

- Add --triplet option to set custom VCPKG triplet. This can be used
to override the above behavior.